refactor(avatars): move F0AvatarPulse out of avatars into sds/Home
F0AvatarPulse is a specific mood/check-in feature (Home DaytimePage and the
RichText MoodTracker), not a base avatar primitive — so it should not live among
the avatar components, and being an sds-level component it should not be
supported inside main-F0 components either.

- Move the component to sds/Home/F0AvatarPulse; regroup its story (and the
  sibling DaytimePage story) under "Home/" in Storybook.
- Extract the shared mood vocabulary (pulses, Pulse, pulseIcon, pulseIconColor)
  to lib/mood so the components-level MoodTracker consumes it without a
  components -> sds layering inversion.
- Update DaytimePage and MoodTracker imports.
- Drop the "pulse" type from F0Card's CardAvatar (and its F0CardRow story/test
  examples): a components-level card must not depend on an sds component. This
  card-avatar type was only exercised by stories/tests, never by product code.

Relocation only — F0AvatarPulse is not stabilized here.
