+++ packages/react/src/components/F0Card/__stories__/F0CardRow.stories.tsx
@@ -275,7 +275,7 @@ export const WithAlert: Story = {
 
 /**
  * The `avatar` prop accepts every single-avatar type in the system — person,
- * company, team, file, flag, icon, emoji, module, alert, date and pulse — each
+ * company, team, file, flag, icon, emoji, module, alert and date — each
  * rendered at a single, fixed size on the left (the size is not configurable).
  */
 export const AvatarTypes: Story = {
@@ -351,18 +351,6 @@ export const AvatarTypes: Story = {
         avatar={{ type: "date", date: new Date(2026, 5, 5) }}
         title="Team offsite"
         description="Date avatar"
-        primaryAction={{ label: "Open", onClick: fn() }}
-      />
-      <F0CardRow
-        avatar={{
-          type: "pulse",
-          firstName: "Jane",
-          lastName: "Cooper",
-          pulse: "positive",
-          onPulseClick: fn(),
-        }}
-        title="Jane Cooper"
-        description="Pulse avatar"
         primaryAction={{ label: "Open", onClick: fn() }}
       />
     </div>
