--- packages/react/src/components/avatars/F0AvatarPulse/index.tsx
@@ -1,2 +1,0 @@
-export { F0AvatarPulse, pulseIcon, pulseIconColor } from "./F0AvatarPulse"
-export type { Pulse } from "./F0AvatarPulse"
