+++ packages/react/src/sds/Home/DaytimePage/index.tsx
@@ -2,7 +2,6 @@ import { cva, type VariantProps } from "cva"
 import { ComponentProps } from "react"
 
 import { F0AvatarPerson } from "@/components/avatars/F0AvatarPerson"
-import { F0AvatarPulse } from "@/components/avatars/F0AvatarPulse"
 import { F0Button } from "@/components/F0Button"
 import { OneSwitch as OnePromotionSwitch } from "@/experimental/AiPromotionChat/OneSwitch"
 import Menu from "@/icons/app/Menu"
@@ -10,6 +9,7 @@ import { withDataTestId } from "@/lib/data-testid"
 import { experimentalComponent } from "@/lib/experimental"
 import { cn } from "@/lib/utils"
 import { useSidebar } from "@/patterns/ApplicationFrame/FrameProvider"
+import { F0AvatarPulse } from "@/sds/Home/F0AvatarPulse"
 import { F0OneSwitch } from "@/sds/ai/F0OneSwitch"
 
 const daytimePageVariants = cva({
