+++ packages/react/src/sds/Home/F0AvatarPulse/index.tsx
@@ -1,0 +1,2 @@
+export { F0AvatarPulse } from "./F0AvatarPulse"
+export type { F0AvatarPulseProps } from "./F0AvatarPulse"
